backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
decorator==5.2.1
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
fsspec==2026.2.0
future==1.0.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huepy==1.2.1
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
instabot==0.117.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mock==5.2.0
moviepy==2.2.1
multidict==6.7.1
numpy==2.4.3
oauth2client==4.1.3
packaging==26.0
pillow==11.3.0
pluggy==1.6.0
postgrest==2.28.2
proglog==0.1.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyDrive==1.3.1
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.3
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
realtime==2.28.2
requests==2.32.5
requests-toolbelt==1.0.0
responses==0.26.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rsa==4.9.1
schedule==1.2.2
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
storage3==2.28.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
tenacity==9.1.4
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
